refactor: split web server bootstrap into setup, runtime, and template modules
